backend/pyproject.toml
[project]
name = "zenn-hackathon-backend"
version = "1.0.0"
description = "Backend API for Zenn Hackathon data management and RAG system"
requires-python = ">=3.11"
dependencies = [
    # Core
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # API
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    # Azure OpenAI
    "openai>=1.12.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-chroma>=0.1.0",
    # Vector Store
    "chromadb>=0.4.22",
    # HTTP Client
    "httpx>=0.26.0",
    # Data Processing
    "pandas>=2.0.0",
    "langchain-text-splitters>=1.1.0",
    "langchain-core>=1.2.0",
    "sse-starlette>=3.0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
